feat: Phase 1 package scaffold — garmin_extract/ package with TUI stub and print menu
- Created garmin_extract/ package (cli.py, app.py, menu.py, __main__.py, __init__.py)
- cli.py: argparse entry point with --no-tui, --dry-run, --verbose, --version
- app.py: Textual TUI stub (Phase 2 placeholder) with auto-detect headless fallback
- menu.py: print menu ported from garmin_extract.py with corrected ROOT path
- garmin_extract.py: converted to backward-compatibility shim
- pyproject.toml: added textual>=0.50.0, [project.scripts], packages.find include
- Fixed datetime.utcfromtimestamp/utcnow deprecation warnings (pullers/garmin.py, reports/build_garmin_csvs.py)
- Updated CLAUDE.md entry points and build phase docs
