+++ garmin_extract/__init__.py
@@ -1,0 +1,8 @@
+"""garmin-extract — automated Garmin Connect data pipeline."""
+
+from importlib.metadata import PackageNotFoundError, version
+
+try:
+    __version__ = version("garmin-extract")
+except PackageNotFoundError:
+    __version__ = "dev"
