+++ garmin_extract/__main__.py
@@ -1,0 +1,4 @@
+from garmin_extract.cli import main
+
+if __name__ == "__main__":
+    main()
