+++ pyproject.toml
@@ -12,7 +12,14 @@ dependencies = [
     "google-auth>=2.0.0",
     "google-auth-oauthlib>=1.0.0",
     "rich>=13.0.0",
+    "textual>=0.50.0",
 ]
+
+[project.scripts]
+garmin-extract = "garmin_extract.cli:main"
+
+[tool.setuptools.packages.find]
+include = ["garmin_extract*"]
 
 [tool.ruff]
 line-length = 100
